+++ src/modules/actor/sheet/cpr-character-sheet.js
@@ -573,12 +573,10 @@ export default class CPRCharacterActorSheet extends CPRActorSheet {
   async _createInventoryItem(event) {
     LOGGER.trace("_createInventoryItem | CPRCharacterActorSheet | Called.");
     const itemType = SystemUtils.GetEventDatum(event, "data-item-type");
-    const itemTypeNice = itemType.toLowerCase().capitalize();
-    const itemString = "ITEM.Type";
-    const itemTypeLocal = itemString.concat(itemTypeNice);
-    const newWord = SystemUtils.Localize("CPR.actorSheets.commonActions.new");
-    const newType = SystemUtils.Localize(itemTypeLocal);
-    const itemName = `${newWord} ${newType}`;
+    const itemString = `TYPES.Item.${itemType}`;
+    const itemName = SystemUtils.Format("CPR.actorSheets.commonActions.new", {
+      item: SystemUtils.Localize(itemString),
+    });
     const itemImage = SystemUtils.GetDefaultImage("Item", itemType);
     const itemData = { img: itemImage, name: itemName, type: itemType };
     await this.actor.createEmbeddedDocuments("Item", [itemData]);